Transactions, Layout: Refactor views
