+++ src/components/sections/index.ts
@@ -1,4 +1,4 @@
 export { default as DebitCardSection } from './debit-card-section.vue';
 export { default as DepositCardSection } from './deposit-card-section.vue';
-export { default as HeaderBalance } from './header-balance.vue';
+export { default as HeaderBalance } from './header-balance-section/header-balance.vue';
 export { default as MenuSection } from './menu-section.vue';
