+++ src/components/transaction-list/index.ts
@@ -1,0 +1,1 @@
+export { default as TransactionList } from './transaction-list.vue';
