--- src/components/wallet-info-rail/index.ts
@@ -1,1 +1,0 @@
-export { default as WalletInfoRail } from './wallet-info-rail.vue';
