--- src/components/wallet-info-rail/transactions-list/dummy-data.ts
@@ -1,30 +1,0 @@
-export const dummyTransactionData: any = [
-  {
-    timeStamp: 1318781876,
-    transactions: [
-      {
-        hash: 'dhdhhdhdhdh',
-        symbol: '0x3fa729b4548becbad4eab6ef18413470e6d5324c',
-        heading: 'Deposit in Savings',
-        amount: '11,984.4953 USDC'
-      },
-      {
-        hash: 'dhdhhddcdschdhdh',
-        symbol: '0x3fa729b4548becbad4eab6ef18413470e6d5324c',
-        heading: 'Withdraw from Savings',
-        amount: '11,984.4953 USDC'
-      }
-    ]
-  },
-  {
-    timeStamp: 1418781276,
-    transactions: [
-      {
-        hash: '13324244',
-        symbol: '0x3fa729b4548becbad4eab6ef18413470e6d5324c',
-        heading: 'Increase Boost',
-        amount: '11,984.4953 USDC'
-      }
-    ]
-  }
-];
