--- src/components/wallet-info-rail/transactions-list/transaction-list.vue
@@ -1,53 +1,0 @@
-<template>
-  <div class="transactions-list">
-    <div v-if="!transactionGroups.length" class="empty-state">
-      {{ $t('lblConnectWalletTransactionHistory') }}
-    </div>
-    <transaction-group
-      v-for="txGroup in transactionGroups"
-      v-else
-      :key="txGroup.date"
-      :heading-text="formatDate(txGroup.timeStamp)"
-      :transactions="txGroup.transactions"
-    />
-  </div>
-</template>
-
-<script lang="ts">
-import Vue from 'vue';
-import dayjs from 'dayjs';
-
-import TransactionGroup from './transaction-group.vue';
-import { mapGetters } from 'vuex';
-
-export default Vue.extend({
-  name: 'TransactionList',
-  components: {
-    TransactionGroup
-  },
-  props: {
-    address: {
-      type: String,
-      required: false
-    }
-  },
-  computed: {
-    ...mapGetters('account', {
-      transactionGroups: 'transactionsGroupedByDay'
-    })
-  },
-  methods: {
-    formatDate(timestamp: number): string {
-      const date = dayjs.unix(timestamp);
-
-      if (dayjs().diff(date, 'days') <= 1) {
-        return date.calendar(undefined, {
-          sameDay: this.$t('dates.sameDay'),
-          lastDay: this.$t('dates.lastDay')
-        });
-      }
-      return date.format('DD MMMM');
-    }
-  }
-});
-</script>
